update cli and api cache

+++ pkg/api/types/v1/request/service.go
@@ -45,7 +45,7 @@ type ServiceOptionsSpec struct {
 	Entrypoint *string        `json:"entrypoint,omitempty"`
 	Command    *string        `json:"command,omitempty"`
 	EnvVars    *[]string      `json:"env,omitempty"`
-	Ports      map[int]string `json:"ports,omitempty"`
+	Ports      map[uint16]string `json:"ports,omitempty"`
 }
 
 // swagger:ignore
